KMP-6: move home tab icons and strings into shared resources
