+++ home/ui/src/androidMain/res/values/strings.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="watch_apps">Watchapps</string>
     <string name="grant">Grant</string>
     <string name="open_settings">Open Settings</string>
     <string name="permissions">Permissions</string>
@@ -24,9 +23,6 @@
     <string name="calendar_permission_title">Calendar</string>
     <string name="calendar_permission_description">Needed to sync calendar to the Timeline</string>
     <string name="developer_connection">Developer connection</string>
-    <string name="watches">Watches</string>
-    <string name="notifications">Notifications</string>
-    <string name="tools">Tools</string>
     <string name="version">MicroPebble version %1$s</string>
     <string name="save_logs">Save logs</string>
     <string name="voice_permission_title">Voice recording</string>
