+++ home/ui/src/commonMain/composeResources/drawable/notifications.xml
@@ -1,7 +1,7 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="24dp" android:tint="#000000"
         android:viewportHeight="960" android:viewportWidth="960" android:width="24dp">
 
-    <path android:fillColor="@android:color/white"
+    <path android:fillColor="#FFFFFFFF"
           android:pathData="M480,880Q447,880 423.5,856.5Q400,833 400,800L560,800Q560,833 536.5,856.5Q513,880 480,880ZM160,760L160,680L240,680L240,400Q240,317 290,252.5Q340,188 420,168L420,140Q420,115 437.5,97.5Q455,80 480,80Q505,80 522.5,97.5Q540,115 540,140L540,168Q551,171 561.5,174.5Q572,178 582,183L519,245Q509,243 500,241.5Q491,240 480,240Q414,240 367,287Q320,334 320,400L320,680L640,680L640,596L720,516L720,680L800,680L800,760L160,760ZM480,520L480,520Q480,520 480,520Q480,520 480,520Q480,520 480,520Q480,520 480,520L480,520L480,520L480,520L480,520L480,520ZM480,520L480,397L701,177Q710,168 721,164Q732,160 743,160Q755,160 766,164.5Q777,169 786,178L823,215Q831,224 835.5,235Q840,246 840,257Q840,268 836,279.5Q832,291 823,300L603,520L480,520ZM780,257L780,257L743,220L743,220L780,257ZM540,460L578,460L699,338L681,319L662,301L540,422L540,460ZM681,319L662,301L662,301L699,338L699,338L681,319Z" />
 
 </vector>
