+++ home/ui/src/commonMain/composeResources/drawable/watches.xml
@@ -1,0 +1,6 @@
+<!-- drawable/watch_variant.xml -->
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="22dp" android:width="22dp"
+        android:viewportWidth="24" android:viewportHeight="24">
+    <path android:fillColor="#000000"
+          android:pathData="M8,0L7.17,5H7A2,2 0 0,0 5,7V17C5,18.11 5.9,19 7,19H7.17L8,24H16L16.83,19H17A2,2 0 0,0 19,17V7C19,5.89 18.1,5 17,5H16.83L16,0H8M7,7H17V17H7V7Z" />
+</vector>
