+++ home/ui/src/commonMain/kotlin/com/matejdro/micropebble/home/HomeScreenContent.kt
@@ -16,17 +16,26 @@ import androidx.compose.material3.NavigationRail
 import androidx.compose.material3.NavigationRailItem
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import com.matejdro.micropebble.home.resources.Res
+import com.matejdro.micropebble.home.resources.notifications
+import com.matejdro.micropebble.home.resources.tools
+import com.matejdro.micropebble.home.resources.watch_apps
+import com.matejdro.micropebble.home.resources.watchapps
+import com.matejdro.micropebble.home.resources.watches
 import com.matejdro.micropebble.navigation.keys.base.SelectedTabContent
 import com.matejdro.micropebble.navigation.keys.base.Tab
+import org.jetbrains.compose.resources.painterResource
+import org.jetbrains.compose.resources.stringResource
 
 @Composable
 fun HomeScreenContent(
-   tabs: List<HomeTab>,
    selectedContent: SelectedTabContent,
    tabletMode: Boolean,
    switchTab: (Tab) -> Unit,
 ) {
+   val tabs = homeTabs()
    val animatedMainContent: @Composable () -> Unit = {
       AnimatedContent(
          selectedContent,
@@ -40,6 +49,30 @@ fun HomeScreenContent(
       NavigationRailContent(animatedMainContent, tabs, selectedContent.tab, switchTab)
    } else {
       NavigationBarContent(animatedMainContent, tabs, selectedContent.tab, switchTab)
+   }
+}
+
+@Composable
+private fun homeTabs(): List<HomeTab> {
+   val watchesIcon = painterResource(Res.drawable.watches)
+   val watchAppsIcon = painterResource(Res.drawable.watchapps)
+   val notificationsIcon = painterResource(Res.drawable.notifications)
+   val toolsIcon = painterResource(Res.drawable.tools)
+   val watchesLabel = stringResource(Res.string.watches)
+   val watchAppsLabel = stringResource(Res.string.watch_apps)
+   val notificationsLabel = stringResource(Res.string.notifications)
+   val toolsLabel = stringResource(Res.string.tools)
+
+   return remember(
+      watchesIcon, watchAppsIcon, notificationsIcon, toolsIcon,
+      watchesLabel, watchAppsLabel, notificationsLabel, toolsLabel,
+   ) {
+      listOf(
+         HomeTab(Tab.WATCHES, watchesIcon, watchesLabel),
+         HomeTab(Tab.WATCH_APPS, watchAppsIcon, watchAppsLabel),
+         HomeTab(Tab.NOTIFICATIONS, notificationsIcon, notificationsLabel),
+         HomeTab(Tab.TOOLS, toolsIcon, toolsLabel),
+      )
    }
 }
